Extract a tidy data frame from RSS and Atom and JSON feeds

tidyfeed() downloads and parses rss feeds. The function produces a tidy data frame, easy to use for further manipulation and analysis.

tidyfeed(feed, sf = TRUE, config = list())

Arguments

feed
(character). The url for the feed that you want to parse.

sf
. If TRUE, returns sf dataframe.

config
Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

References

https://en.wikipedia.org/wiki/RSS

Examples

# NOT RUN {
# Atom feed:
tidyfeed("http://journal.r-project.org/rss.atom")
# rss/xml:
tidyfeed("http://fivethirtyeight.com/all/feed")
# jsonfeed:
tidyfeed("https://daringfireball.net/feeds/json")
# georss:
tidyfeed("http://www.geonames.org/recent-changes.xml")
# }